i2c: add soft stop option
When we call i2cStop() we are aborting the current transfer,
disabling DMA, disabling interrupts and then cutting the clock to the
peripheral.

If i2cStop() is called after each transfer, for example to be able
to share DMA, this has the nasty consequence of breaking the stop
condition and NACK'ing on received bytes from a slave. This happens
because when the perifpheral didn't have time yet to complete the stop
condition and we just go ahead and disable everything

This provides a "soft stop" option, in which we disable DMA, disable
interrupts, but we keep the peripheral turned on: this is sufficient
for making sure the peripheral won't trigger any transaction due to
noise in the bus and that the DMA won't be used. This way stop
condition and acks work again when calling i2cSoftStop() after each
i2c transaction.

(cherry picked from commit 540364a3c02e4713b7383ef070cd6419721a36eb)